Adding pp2 exercises 1 and 2 to my pp2 lab folder
I'm not sure if my work subdirectory will also be included in this commit. When I used git add, I added the entire pp2 folder, but I'm not sure if this only tracked the files directly listed in that folder. I'm using this commit to find out! Also, I am using this commit to test out the commit subject versus body message.